+++ frykit/_artist.py
@@ -314,6 +314,7 @@ def rectangle_path(x0: float, x1: float, y0: float, y1: float) -> Path:
     return path
 
 
+# TODO: https://github.com/SciTools/cartopy/issues/1830
 class Frame(Artist):
     '''GMT风格的边框'''