+++ script/run_example.py
@@ -6,6 +6,8 @@ from pathlib import Path
 
 code_filepath = Path(__file__).absolute()
 example_dirpath = code_filepath.parent.parent / 'example'
-for example_filepath in example_dirpath.iterdir():
-    subprocess.run(['python', str(example_filepath)])
-    print(f'> {example_filepath.name}')
+for filepath in example_dirpath.iterdir():
+    if filepath.stem == 'clabel':
+        continue
+    subprocess.run(['python', str(filepath)])
+    print(f'> {filepath.name}')